+++ noodles-csi/src/writer.rs
@@ -5,8 +5,9 @@ use noodles_bgzf as bgzf;
 
 use super::{
     index::{
+        header::ReferenceSequenceNames,
         reference_sequence::{bin::Chunk, Bin, Metadata},
-        ReferenceSequence,
+        Header, ReferenceSequence,
     },
     BinningIndex, Index, MAGIC_NUMBER,
 };
@@ -58,7 +59,7 @@ where
         let depth = i32::from(index.depth());
         self.inner.write_i32::<LittleEndian>(depth)?;
 
-        write_aux(&mut self.inner, index.aux())?;
+        write_aux(&mut self.inner, index.header())?;
         write_reference_sequences(&mut self.inner, index.depth(), index.reference_sequences())?;
 
         if let Some(n_no_coor) = index.unplaced_unmapped_record_count() {
@@ -76,15 +77,78 @@ where
     writer.write_all(MAGIC_NUMBER)
 }
 
-fn write_aux<W>(writer: &mut W, aux: &[u8]) -> io::Result<()>
+fn write_aux<W>(writer: &mut W, header: Option<&Header>) -> io::Result<()>
 where
     W: Write,
 {
+    let mut aux = Vec::new();
+
+    if let Some(hdr) = header {
+        write_header(&mut aux, hdr)?;
+    }
+
     let l_aux =
         i32::try_from(aux.len()).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_i32::<LittleEndian>(l_aux)?;
 
-    writer.write_all(aux)?;
+    writer.write_all(&aux)?;
+
+    Ok(())
+}
+
+pub(crate) fn write_header<W>(writer: &mut W, header: &Header) -> io::Result<()>
+where
+    W: Write,
+{
+    let format = i32::from(header.format());
+    writer.write_i32::<LittleEndian>(format)?;
+
+    let col_seq = i32::try_from(header.reference_sequence_name_index())
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    writer.write_i32::<LittleEndian>(col_seq)?;
+
+    let col_beg = i32::try_from(header.start_position_index())
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    writer.write_i32::<LittleEndian>(col_beg)?;
+
+    let col_end = header.end_position_index().map_or(Ok(0), |i| {
+        i32::try_from(i).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
+    })?;
+    writer.write_i32::<LittleEndian>(col_end)?;
+
+    let meta = i32::from(header.line_comment_prefix());
+    writer.write_i32::<LittleEndian>(meta)?;
+
+    let skip = i32::try_from(header.line_skip_count())
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    writer.write_i32::<LittleEndian>(skip)?;
+
+    write_reference_sequence_names(writer, header.reference_sequence_names())?;
+
+    Ok(())
+}
+
+fn write_reference_sequence_names<W>(
+    writer: &mut W,
+    reference_sequence_names: &ReferenceSequenceNames,
+) -> io::Result<()>
+where
+    W: Write,
+{
+    const NUL: u8 = 0x00;
+
+    // Add 1 for each trailing NUL.
+    let len = reference_sequence_names
+        .iter()
+        .map(|n| n.len() + 1)
+        .sum::<usize>();
+    let l_nm = i32::try_from(len).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+    writer.write_i32::<LittleEndian>(l_nm)?;
+
+    for reference_sequence_name in reference_sequence_names {
+        writer.write_all(reference_sequence_name.as_bytes())?;
+        writer.write_u8(NUL)?;
+    }
 
     Ok(())
 }